ditch table-ann, add support for pseudo-expression ALL to table instead
