+++ tests/testcases/table/test_ann_all/config.yaml
@@ -1,0 +1,2 @@
+function: "table"
+expression: "ALL"
