+++ tests/testcases/table/test_ann_exon_range/config.yaml
@@ -1,2 +1,3 @@
 function: "table"
 expression: "CHROM, POS, ANN['EXON']"
+wide: []
