+++ tests/testcases/table/test_ann_exon_range_undefined/config.yaml
@@ -1,2 +1,3 @@
 function: "table"
 expression: "CHROM, POS, ANN['Protein_position'].start, ANN['Protein_position'].length, ANN['Protein_position']"
+wide: []
