+++ tests/testcases/table/test_ann_test/config.yaml
@@ -1,2 +1,3 @@
 function: "table"
 expression: "INDEX, CHROM, POS, ANN['Feature_ID']"
+wide: []
