+++ tests/testcases/table/test_ast/config.yaml
@@ -1,2 +1,3 @@
 function: "table"
 expression: (POS, POS)
+wide: []
