+++ tests/testcases/table/test_chrom_qual/config.yaml
@@ -1,2 +1,3 @@
 function: "table"
 expression: "CHROM, QUAL"
+wide: []
