+++ tests/testcases/table/test_default_get/config.yaml
@@ -1,2 +1,3 @@
 function: "table"
 expression: "CHROM, POS, ANN.get('IMPACT', 'DEFAULT'), ANN.get('IMPACT', NA), ANN.get('IMPACT'), ANN['IMPACT'], INFO.get('SVLEN', 0)"
+wide: []
