+++ tests/testcases/table/test_format_all/test.vcf
