+++ tests/testcases/table/test_header_col_number_test/config.yaml
@@ -2,3 +2,4 @@ function: "table"
 expression: "CHROM, POS, for_each_sample( lambda sample: FORMAT['GT'][sample] )"
 header: "for_each_sample(lambda sample: f'useless_concat_on_sample_{sample}')"
 raises: "HeaderWrongColumnNumberError"
+wide: []
