+++ tests/testcases/table/test_header_none_test/config.yaml
@@ -1,3 +1,4 @@
 function: "table"
 expression: "CHROM, QUAL"
 header: "none"
+wide: []
