+++ tests/testcases/table/test_sample_list_comprehension_test/config.yaml
@@ -1,3 +1,4 @@
 function: "table"
 expression: "CHROM, [ FORMAT['DP'][sample] for sample in SAMPLES ]"
 header: "auto"
+wide: []
