+++ tests/testcases/table/test_sample_long/config.yaml
@@ -1,4 +1,3 @@
 function: "table"
 expression: "CHROM, round(FORMAT['AD'][SAMPLE][0] / FORMAT['DP'][SAMPLE] * INFO['MQ'])"
 header: "auto"
-long: []
