+++ tests/testcases/table/test_sample_long_header/config.yaml
@@ -1,4 +1,3 @@
 function: "table"
 expression: "CHROM, round(FORMAT['AD'][SAMPLE][0] / FORMAT['DP'][SAMPLE] * INFO['MQ'])"
 header: "CHROM, round(AD/DP*MQ)"
-long: []
