+++ tests/testcases/table/test_separator_test/config.yaml
@@ -1,3 +1,4 @@
 function: "table"
 expression: "CHROM, QUAL"
 separator: "#"
+wide: []
