+++ tests/testcases/table/test_stop/config.yaml
@@ -1,2 +1,3 @@
 function: "table"
 expression: "POS, END"
+wide: []
