--- tests/testcases/table_all/test_ann/config.yaml
@@ -1,1 +1,0 @@
-function: "table-all"
