--- tests/testcases/table_all/test_format/config.yaml
@@ -1,1 +1,0 @@
-function: "table-all"
